Speed up SAM transcoder
Currently, the SAM transoding works in three steps:

1. Convert SAM records with same query name to MPEGG records and
save them to files accoring to their reference sequence
2. Split each file into chunks and sort them
3. Perform a merge sort of all files

However it is possible to perform steps one and two at once by extending the
comparator and sorting not only by position but simultaniously by
reference sequence.

This way a complete write to disk is avoided. Speedup is around 40%.