Multithreading SAM transcoder
Enables multithreading support for the SAM transcoder. Conversion
of SAM to MPEGG records and all of the sorting can happen in parallel.
Only the read operation must be synchronized. Write can be parallelized
as the outputs are written to separate files per chunk.

An 80% speedup can be observed when using 8 threads instead of one.
I/O speed remains the limiting factor.