Bypass bitwriter for speedup
This is valid, as in the MPEGG records, all fields are byte aligned.
The sequence and quality value fields are the largest fields and
saving the overhead of bitfiddling speeds up SAM to MPEGG conversion
by around 35%.

Enabling bypass for all other fields increaes performance by
a few more percent.